Break up group-ban/kick tests into separate files.
