+++ apps/tlon-web/e2e/group-ban-core.spec.ts
@@ -1,0 +1,114 @@
+import { expect } from '@playwright/test';
+
+import * as helpers from './helpers';
+import { test } from './test-fixtures';
+
+test('should allow banning and unbanning users in public groups', async ({
+  zodSetup,
+  tenSetup,
+}) => {
+  const zodPage = zodSetup.page;
+  const tenPage = tenSetup.page;
+
+  // Create a public group
+  await helpers.createGroup(zodPage);
+
+  // Keep it public (default)
+  await helpers.openGroupSettings(zodPage);
+  // Just verify we're in settings, don't check exact text
+  await expect(zodPage.getByText('Group info')).toBeVisible();
+
+  // Invite ~ten
+  await zodPage.getByText('Invite people').click();
+  await zodPage.getByPlaceholder('Filter by nickname, @p').fill('~ten');
+  await zodPage.waitForTimeout(1000);
+  await zodPage.getByTestId('ContactRow').getByText('~ten').first().click();
+  await zodPage.getByText('Invite 1 and continue').click();
+  await zodPage.waitForTimeout(1000);
+
+  // ~ten accepts invite
+  await helpers.acceptGroupInvite(tenPage);
+
+  // Navigate to members page and verify ban option exists
+  await helpers.openGroupSettings(zodPage);
+  await zodPage.getByTestId('GroupMembers').click();
+  await zodPage.waitForTimeout(2000);
+
+  // Click on ~ten to open profile sheet
+  await zodPage.getByTestId('MemberRow').filter({ hasText: '~ten' }).click();
+
+  // Verify ban option is available and click it
+  await expect(zodPage.getByText('Ban User')).toBeVisible();
+  await zodPage.getByText('Ban User').click();
+
+  // Wait for ban to complete
+  await zodPage.waitForTimeout(3000);
+
+  // Close sheet if needed
+  await zodPage.keyboard.press('Escape');
+  await zodPage.waitForTimeout(1000);
+
+  // Verify that ~ten no longer appears in regular members
+  // Note: They should now be in banned users section, but we'll just verify they're not in regular list
+  const memberRows = zodPage.getByTestId('MemberRow');
+  const tenRow = memberRows.filter({ hasText: '~ten' });
+
+  // The member might still be visible in banned section, so we just verify ban action worked
+  // by checking that we can now see unban option if we click on them again
+  const bannedSection = zodPage.getByText('Banned Users');
+  await expect(bannedSection).toBeVisible({ timeout: 5000 });
+
+  // Click on ~ten in the banned section
+  await tenRow.click();
+  await expect(zodPage.getByText('Unban User')).toBeVisible();
+
+  // Verify ~ten can no longer see the group after being banned
+  // Navigate to Home on ~ten's page
+  await tenPage.getByTestId('HomeNavIcon').click();
+  await tenPage.waitForTimeout(2000);
+
+  // The group should not be visible in ~ten's chat list
+  await expect(
+    tenPage.getByTestId('ChatListItem-Untitled group-unpinned')
+  ).not.toBeVisible({ timeout: 5000 });
+
+  // Now unban ~ten
+  await zodPage.getByText('Unban User').click();
+  await zodPage.waitForTimeout(3000);
+
+  // Close sheet
+  await zodPage.keyboard.press('Escape');
+  await zodPage.waitForTimeout(1000);
+
+  // Verify ~ten is no longer in the banned section
+  // After unbanning, the banned section should disappear or ~ten should not be in it
+  await expect(zodPage.getByText('Banned Users')).not.toBeVisible({
+    timeout: 5000,
+  });
+
+  // After unbanning, ~ten still can't see the group (needs re-invite)
+  // This verifies that unbanning doesn't automatically restore access
+  await expect(
+    tenPage.getByTestId('ChatListItem-Untitled group-unpinned')
+  ).not.toBeVisible({ timeout: 5000 });
+
+  // Navigate to group settings to re-invite
+  await helpers.openGroupSettings(zodPage);
+
+  // Re-invite ~ten to the group (now that they've been unbanned)
+  await zodPage.getByText('Invite people').click();
+  await zodPage.getByPlaceholder('Filter by nickname, @p').fill('~ten');
+  await zodPage.waitForTimeout(1000);
+  await zodPage.getByTestId('ContactRow').getByText('~ten').first().click();
+  await zodPage.getByText('Invite 1 and continue').click();
+  await zodPage.waitForTimeout(3000); // Wait for invite to propagate
+
+  // Refresh ~ten's page to ensure invite appears
+  await tenPage.reload();
+  await tenPage.waitForTimeout(2000);
+
+  // ~ten accepts re-invite
+  await helpers.acceptGroupInvite(tenPage);
+
+  // Verify ~ten can access the group again after being unbanned and re-invited
+});
